+++ server/index.js
@@ -1,9 +1,11 @@
 const express = require('express');
 const path = require('path');
 const passport = require('passport');
+const bodyParser = require('body-parser');
 
 const app = express();
-var bodyParser = require('body-parser');
+const api = express.Router();
+const user = express.Router();
 
 app.use(express.static(path.resolve(__dirname, '..', 'build')));
 
@@ -11,12 +13,17 @@ app.get('/', (req, res) => {
   res.sendFile(path.resolve(__dirname, '..', 'build', 'index.html'));
 });
 
-app.listen(process.env.PORT || 8080);
-
+app.use('/api', api);
+app.use('/user', passport.authenticate('jwt', { session: false }), user);
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(passport.initialize());
 
-module.exports = app;
+app.listen(process.env.PORT || 8080);
+
+module.exports = { app, api, user };
 
 require('./samples.js');
-require('./routes/routes.js');
+
+require('./routes/auth.js');
+require('./routes/user.js');
+require('./routes/books.js');
