+++ server/models/users.js
@@ -1,5 +1,4 @@
-const mongoose = require('./mongoose.js');
-const Schema = mongoose.Schema;
+const { mongoose, Schema } = require('./mongoose.js');
 const bcrypt = require('bcrypt');
 
 const usersSchema = new Schema({
@@ -10,10 +9,12 @@ const usersSchema = new Schema({
 usersSchema.pre('save', function(next) {
   if (this.isNew) {
     const saltRounds = 10;
+
     bcrypt.hash(this.password, saltRounds, (err, hashedPassword) => {
       if (err) {
         return next(err);
       }
+
       this.password = hashedPassword;
       next();
     });
