+++ server/routes/books.js
@@ -1,4 +1,4 @@
-const { api } = require('./routes.js');
+const { api } = require('../index.js');
 const Books = require('../models/books.js');
 
 api.get('/books', (req, res) => {
