+++ server/routes/user.js
@@ -1,4 +1,4 @@
-const { user } = require('./routes');
+const { user } = require('../index.js');
 
 user.get('/profile', function(req, res) {
   res.send(req.user);
